backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
supabase==2.11.0
httpx==0.27.2
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.20
